Rename enums to globals, add MAX_INT to globals
