+++ copis/classes/settings.py
@@ -17,7 +17,7 @@ from dataclasses import dataclass
 from typing import List, Dict
 from glm import vec3
 
-from copis.enums import DebugEnv
+from copis.globals import DebugEnv
 from copis.helpers import Point5
 from . import Device, Chamber, BoundingBox
